Add interactive Clojure editor component
An interactive Clojure editor is now available based on Codemirror for syntax
highlighting and sci (small clojure interpreter) for safe evaluation of
Clojure in the browser.

The component is implemented as a function using react hooks to hold
local state. It uses debouncing to evaluate the users code at most every
500 milliseconds.

Both the input code and the result is syntax highlighted.

The component should make it possible to create interactive Clojure blog
posts about Clojure/lisp syntax.

Fix #4.