Compile clojure code and place in static folder
The previous setup where compiled ClojureScript code was imported
directly in the JavaScript files made the webpack bundling process very
slow and resource hungry.

The GitHub action that was doing the gatsby build ran out of memory.

This commit tries a different approach where the ClojureScript is
compiled to a standaloen file in the static folder. Files in the static
folder is not touched by webpack.

The file is imported runtime using react hooks and the imperative dom
api. Only when the dom api is finished loading the script does the lazy
clojure snippet component wrapper render the clojure editor which
references functions set by the now loaded compiled clojurescript.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>