Limit sci to 100 realizations to avoid inifinite evaluation
Sci will by default happily consume inifinte sequences forever. In a
browser context where sci runs in the render thread this is bad news for
the user. To avoid this we can use the max-realizations option to limit
sci to a certain number of realizations over which it will return an
error.

Fix #6.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>