Capture prints from clojure snippets and print
By default no side-effecting functions are available in sci. So a user
printing to *out* will not see any of her printouts in the result
window. To make print style debugging possible we bind sci's out to
goog's string builder. Any time a user prints inside a clojure snippet
it will be added to the string builder.

When the evaluation is done it is returned together with the string in
the string builder, both returned inside a javascript object consumed in
the ClojureEditor react component.

Fix #5.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>