Use gatsby filesystem to import compiled cljs
Placing compiled ClojureScript files inside the static folder manually proved
to be problematic as we then don't get content hashes in the filename
preventing cache busting.

A better approach is using gatsby-source-filesystem to import all files
in the cljs folder into the project. The gatsby plugin takes care of
hashing the file contents and has handy graphql properties to get the
relative URL of the file.

We then just have to run a graphql query inside the clojure snippet
component which gets the url and then uses the existing react effects
and state logic to fetch the javascript file and embedding it in the
html head element.

Fix #7

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>