Add rss feed via gatsby-plugin-feed
RSS is a good format to allow people to subscribe to a blog. It is open
and there are many readers supporting RSS. This commit adds an RSS feed
using the default gatsby feed plugin. Interactive content will not be
visible in the feed, but most static html content should render ok.

Fix #1

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>