Remove images with relative urls
The images from the old blog was not part of the backup. Further more
refering to relative urls in links is not allowed in rss feeds. To fix
the issue I simply removed the images from the posts.

Going forward images should be imported using gatsby file system and
will be given full, valid, urls.

Should fix #8