Use StaticQuery wrapper to fix faulty useStaticQuery
The useStaticQuery hook in the clojure-snippet module seems to be broken
for some reason. Maybe because the component is imported via an mdx
file. In any case the query won't run when running gatsby build, even
though it works when running gatsby develop.

A workaround seems to be using the higher order StaticQuery component
instead which serves much the same function.

Signed-off-by: Snorre Magnus Davøen <snorre.magnus.davoen@gofjords.com>