Improve thread and reply views
New utility functions for creating and flattening the threads.
Normalises into flat list with UI properties for whether to display
parent and child reply lines. Loosly based on the Bsky App code for
thread handling, but simplified for my own needs.

Reply functionality is now opened in a dialog element. This allows us to
show a modal style editor while the commenter is replying. Need to fix
the dialog state handling as it is quite buggy now.

Introduced light and dark mode class names in the comment system
components. All components now support both modes.

The text limit is 300 characters for a Bsky post so increasing counter
to reflect this.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>