feat: Comments system built on atproto and bluesky (#28)
A comment system built on top of AT Protocol and the Bluesky social network.
The system is implemented as a client side Astro island that interacts directly
with the Bluesky server. It uses Solid.js library for rendering the view and the
same Tailwind setup as the rest of the blog for styling.

Auth is handled client side and the refresh token stored in local storage to allow
resumability of the session.

The implementation borrows the flatten logic from the Bsky app to produce a
similar ordering of posts to Bsky. It does not however do any of the sorting that
Bsky does.

Currently it is not possible to highlight a post down in the thread, but this feature is not
too far off from an implementation setup as most of the code should support fetching
and displaying a thread where the highlighted post starts further down in the thread.

A blog post goes into some details about why and how I came to implement the
ATProto/Bluesky comment system.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>