feat: View transitions on page navigation
Enable the experimental view transitions support. View transitions are a
new web feature, currently supported only in Chromium, that enables the
browser to keep around dom elements when doing page navigations. This
makes the page look smoother as the user navigates the page.

It also adds support to control how dom elements moves on the page. For
example you could morph a title from one page by associating each title
with eachother via a transition name. Then a h2 in a list could morph up
to become the h1 of the page the user navigates to. Currently I do not
use this feature.

It was tested on the blog title in the blog post list and on the blog
pages. This looked good when navigating to a blog post page, but looked
less than ideal when navigating back to the blog post list.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>