feat: Support continue thread and highlight post (#40)
There is a limit to how many children are fetched using the Bluesky
client. Whenever we reach the limit, i.e. the reply count indicates more
replies, but the reply list is empty, we show a "continue thread" button
to allow readers to show more comments. This highlights the last reply
and shows replies under that post.

Each post header (avatar, display name, etc) is a button which allows
users to highlight that post. This is useful if you want to reduce the
part of the thread tree that you are looking at.

Note that the parentHeight limit is now set to 20. This makes it so that
up to 20 parent posts are fetched allowing readers to view most of the
parents for a highlighted post. In extreme cases where the depth of a
thread is more than 20 some parents are not returned. This is
acceptable.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>