Add recipes and working recipe downloader script
The babashka script for downloading recipes now works. It fetches new
recipes after any already downloaded. It downloads recipes and writes
them with pretty-printed JSON to the beer-recipes folder. Each file name
includes the date at which the recipe was published and the name. This
should give reasonably good ordering of the files.

The script uses the 1password command line utility to fetch the
Brewfather API secret. This allows us to keep the secret protected at
all time, not exposing it to the system unencrypted.

Recipe files includes in source repository so they only have to be
downloaded once.