Implement astro pages for beers list and beer recipe
Adds new Astro pages for the beer recipe list and the beer recipe pages.
Exposes the beer recipe json files as a typed and zod validated Astro
collection so it is easy to work with. Only typed as detailed as the
information needed in the UI as the beer model in Brewfather is quite
large.

Moved beer recipes into folder beers under the content folder as
Astro requires collections to be stored in a specific folder. Download
script updated accordingly.

Currently the recipe layout is very simple and minimalist. I will likely
add more information in a alter commit.