feat: Add beer recipes from brewfather (#50)
* Add recipes and working recipe downloader script

The babashka script for downloading recipes now works. It fetches new
recipes after any already downloaded. It downloads recipes and writes
them with pretty-printed JSON to the beer-recipes folder. Each file name
includes the date at which the recipe was published and the name. This
should give reasonably good ordering of the files.

The script uses the 1password command line utility to fetch the
Brewfather API secret. This allows us to keep the secret protected at
all time, not exposing it to the system unencrypted.

Recipe files includes in source repository so they only have to be
downloaded once.

* Put bb script into separate scripts folder

To keep the project clean, keep helper scripts in a scripts folder.

* Implement astro pages for beers list and beer recipe

Adds new Astro pages for the beer recipe list and the beer recipe pages.
Exposes the beer recipe json files as a typed and zod validated Astro
collection so it is easy to work with. Only typed as detailed as the
information needed in the UI as the beer model in Brewfather is quite
large.

Moved beer recipes into folder beers under the content folder as
Astro requires collections to be stored in a specific folder. Download
script updated accordingly.

Currently the recipe layout is very simple and minimalist. I will likely
add more information in a alter commit.

* Add beer recipe link to menu