feat: Beer glass color indicators
Add new beer glass, tulip shape, to the beers and beer recipe pages.
The beer glass icon adapts to the color of the beer so visitors get a
visual representation of how light or dark the beer is. The color scheme
is based on an SRM to RGB lookup table kindly provided by Torstein
Thune.

The commit also provides som fixes for mobile view so things don't look
so squished.