feat: Blog post likes (#53)
Implement a blog posts like feature using Cloudflare D1 SQLite database,
Drizzle ORM, Astro Cloudflare adapter, and Astro API page routes.

Use Drizzle as migrator, schema creator, and query builder to handle
likes in a database. Make Astro route files with API endpoints (GET and
POST) for likes. Each like row in the DB has two columns, the slug of
the post for which the like is added, and sessionId which is the uuid
identifying a visitor via a cookie.

Add info about the technical session cookie in the privacy policy.
Considering the cookie is not used to track the user or to handle
personal information a cookie banner should not be necessary. Worst case
a dialog can be added to the like button asking for concent, but this is
likely overkill.