refactor: Make projects content collection
Move projects file to a content collection to follow the new best
practices. Render all projects in the main projects.astro page component
as we do not need sub pages for each project. At least for now.

Use rss schema to potentially allow for exposing projects list as rss in
the future.