fix: Content jumps because of inconsistent scrollbar
Because some pages have scrollbars and some don't the content appear to
jump. This is especially noticeable because the site uses browser native
transitions to not re-render the entire page. Because there is no
visible reload flash, the content can be observed more directly jumping.

Added overflow-y scroll to the html element to always show a scrollbar.
This ensures the viewport is the same width on all pages.