feat: Likes on content list pages
Implement support for like counters on content list pages like the blog
list page. In order for this to be performant we want to avoid running
individual http requests per blog, recording or beer. While there are
lots of included requests in Cloudflare Pages, it is good practice not
to use them unecessarily.

Move from Drizzle to kysely. Drizzle proved to have some issues
generating correct migrations from the TypeScript schema definitions.
For example it did not add alter table statements when I added not null
to a schema property (table column).

This yet again shows that native SQL schema first is the best approach.
Kysely allows me to write the SQL schemas as its creators intended, and
then generate TypeScript types for them afterwards.

Note that Kysely does not transform values out of the box, so for epoch
times and such I need to manually transform where needed.