fix: Likes not fetched on collection pages (#66)
Imlement a very simple fetch function with a Promise cache that caches
the fetch response Promise of each collection to avoid duplicated
fetches. If collection is not in the cache we make a new fetch and
insert the returned Promise into the cache before returning to the
caller which can await it. Otherwise we return the cached Promise at
once.